Improved Geohash calculations.

+++ src/main/index.js
@@ -385,10 +385,10 @@ layerControl.addOverlay(new GridCoords(), "Grid cells");
 
 if (location.hash) {
   let coords = location.hash.split(",");
-  let latitude = parseInt(coords[0].slice(1));
-  let longitude = parseInt(coords[1]);
+  let latitude = parseFloat(coords[0].slice(1));
+  let longitude = parseFloat(coords[1]);
   if (coords[2]) {
-    let scale = parseInt(coords[2].split(/^(\d*)z$/g)[1]);
+    let scale = parseFloat(coords[2].split(/^(\d*)z$/g)[1]);
     map.setView([latitude, longitude], scale);
   } else {
     map.setView([latitude, longitude]);
